Make sure to handle std::env::current_exe differences per platform
On macOS the `std::env::current_exe()` path typically does not get
resolved (so we can assert that the current_exe is actually within the
symlinked path), but on Linux the `current_exe()` result is usually
canonicalized (such that the symlink is no longer part of the path, but
it is instead the "real" bin path).

The rest of our logic doesn't actually care if we are seeing a symlink
or not, we are just trying to confirm the executable works.