ci: Ensure all tests run in CI
Without the `--no-fail-fast` option `cargo nextest run` will stop
running tests as soon as it hits the first failure. That's a great
feature, but out whole test suite doesn't take too long and it's much
better to see all of the failures in CI rather than just the first one.