fix(cache-shell-startup): Allow leading whitespace before pragmas
The code prior to this **only** supported `# CMD:` if it was at the very
beginning of a line, but sometimes you might want to put `# CMD: ` inside a
conditional (where you'd usually you'd indent the branches).

Consider this:

```bash
if [ -f "filename.txt" ]; then
  # CMD: echo hello
fi
```

After these changes, that works properly.